rename kernel + merge benchmark and test
